#409 Move definition of value clipper
