+++ neclib/core/__init__.py
@@ -9,6 +9,7 @@ since such dependencies would create circular imports.
 
 from . import environ  # noqa: F401
 from . import formatting  # noqa: F401
+from . import logic  # noqa: F401
 from . import type_aliases  # noqa: F401
 from . import units  # noqa: F401
 from .configuration import *  # noqa: F401, F403
