+++ neclib/core/logic/__init__.py
@@ -1,0 +1,3 @@
+from .clip import clip
+
+__all__ = ["clip"]
